% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Find the euclidean distance between two points}
\usage{
euclidean_distance(pts)
}
\arguments{
\item{pts}{2x2 matrix with x,y values in each row}
}
\value{
The euclidean distance between the points
}
\description{
Find the euclidean distance between two points
}
