% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_distance.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Find the Manhattan distance between two points}
\usage{
manhattan_distance(pts)
}
\arguments{
\item{pts}{2x2 matrix with x,y values in each row}
}
\value{
The Manhattan distance between the points
}
\description{
Find the Manhattan distance between two points
}
