% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minkowski_distance.R
\name{minkowski_distance}
\alias{minkowski_distance}
\title{Find the Minkowski distance between two points.}
\usage{
minkowski_distance(pts, p)
}
\arguments{
\item{pts}{2x2 matrix with x,y values in each row}

\item{p}{The order of the distance.}
}
\value{
The Minkowski distance between the points
}
\description{
See \url{https://en.wikipedia.org/wiki/Minkowski_distance}.
}
